style: Add borders and text around QandaA containers
Currently, there is no visual clue for the users as to where both the
question and answer start and end. This commit adds some borders and
text labels around the question and the answer to help visually identify
what part is the question and what part is the answer.

Before this commit, the question and answer look like this:

    How to see what is in gl/devel but not in HEAD?

    Easy enough: 😄

      $ git log HEAD..gl/devel

And this is how it should like after this commit:

                          +----------+
    +---------------------| question |-+
    |                     +----------+ |
    | How to see what is in gl/devel   |
    | but not in HEAD?                 |
    +----------------------------------+
                            +--------+
    +-----------------------| answer |-+
    |                       +--------+ |
    | Easy enough: 😄                  |
    |                                  |
    |   $ git log HEAD..gl/devel       |
    +----------------------------------+