chore: Fix some reST incorrect syntax
Currently, there something like this:

    See this example::

    .. code-block::

       code example here

Note the double colon `::` preceding the code block. That is invalid
and yields wrong rendering. Either don't use the double colon or don't
use the explicit `.. code-block::` markup.

Also reorder the use of the link directive to make it closer together
with the place where it is used.