style: Use normal fonts for comments
Italic font in comments hurt readability. Not all comments are composed
of alphanumeric-only characters, and a lot of non-alphanumeric
characters look ugly and unreadable with italic fonts, not to mention
that some fonts make those italic chars look even worse.

This commit adds a line of CSS to override pygments and make comment
fonts normal instead of the default italic.