chore: Add CC BY-SA 3.0 Unported licensing
Since the content of this entire project is partly my creation, things I
think about, play with and write about, but also notes and explanation
on other people's works (like books, courses and tutorials), it seems
correct to license this work under the “Creative Commons
Attribution-ShareAlike 3.0 Unported License.”

As a side note, many people push solutions to books, courses, etc. on
public repositories. I'm doing the same except it is also rendered on a
website, something like a public wiki of sorts.

REFERENCES:

• https://creativecommons.org/licenses/by-sa/3.0/
• https://creativecommons.org/2011/04/15/plaintext-versions-of-creative-commons-licenses-and-cc0