chore: Disable next/prev buttons at the bottom
Since this is not a normal documentation site, but rather a wiki-like
thing, not sure if it makes sense to have prev/next navigation links. It
is strange that after "javascript decorations" the next link is
"compiling chicken scheme" 😅. Let's use this if False thing to disable
them for now without messing too much with the theme.

It would be useful for my books and video solutions, but then I would
have to find a way to enable it for certain thigns or pages only.
Perhaps there is a way.