chore: Add ToC on home page body
The site already has a left navigation with links to the topics
published so far. This commit also adds a ToC on the home page body,
after the introductory text and license that shows the main headings for
the topics published so far so we have a more contentful home page.