chore: Override base template
Add the ability to set the <title> value through the front matter. If it
is there, use it as the browser title and use the h1 heading title for
the document display, otherwise, just use the h1 heading.

Also add the ability to set the meta description.

On front matter of any `md` (myst) file, do something like this:

```
---
title: Event Horizon | Universe
description: In this document we shall discuss the wonders of the
universe, and how the Even Horizon breeds Quantum Butterflies.
---
```

The Browser would then say:

> Even Horizon | Universe | Dev How To

And the meta description should also be there on the page HTML helping
SEO and users to understand the contents of the page.