chore: Prevent page scroll when clicking left nav item
A few commits ago, we introduced a bit of JS to cause the current active
navigation item on the left bar to scroll into view. Sadly, it also
started causing the main content to scroll down too. This commit changes
the settings for `scrollIntoView()` function in a way that should stop
the main content to scroll.